refactoring: subpackage for data classes
