+++ proj_eclipse/src/main/java/smartmag/data/Box.java
@@ -1,4 +1,4 @@
-package smartmag;
+package smartmag.data;
 
 public class Box {
