+++ proj_eclipse/src/main/java/smartmag/data/Movimentazione.java
@@ -1,4 +1,4 @@
-package smartmag;
+package smartmag.data;
 
 import java.util.Objects;
