+++ proj_eclipse/src/main/java/smartmag/data/StatoMovimentazione.java
@@ -1,4 +1,4 @@
-package smartmag;
+package smartmag.data;
 
 public enum StatoMovimentazione {
 	NON_ASSEGNATA, PRESA_IN_CARICO, PRELEVATA, COMPLETATA, ANNULLATA;
