+++ proj_eclipse/src/main/java/smartmag/data/StatoOrdine.java
@@ -1,4 +1,4 @@
-package smartmag;
+package smartmag.data;
 
 /**
  * Stato ordine
