+++ proj_eclipse/src/main/java/smartmag/data/TipoUtente.java
@@ -1,4 +1,4 @@
-package smartmag;
+package smartmag.data;
 
 /**
  * Rappresenta i tipi di utenti presenti in magazzino
