+++ proj_eclipse/src/main/java/smartmag/data/Utente.java
@@ -1,4 +1,4 @@
-package smartmag;
+package smartmag.data;
 
 import java.util.EnumSet;
 import java.util.Objects;
