+++ proj_eclipse/src/test/java/smartmag/ProdottoTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import smartmag.data.Prodotto;
+
 class ProdottoTest {
 
 	@Test
